+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 ### Added
+- Add `IdentificationEndpoint.delete`
 - Add `IdentifyScenarioVersion.expireAfterHours`
 - Add `IdentifyScenarioVersion.discardCompletedAfterDays`
 - Add AccountSecurity.discardedIdentificationRetention